update props for sitesearchlistcontainer
